+++ src/org/pipservices3/components/Component.java
@@ -5,6 +5,7 @@ import org.pipservices3.components.count.CompositeCounters;
 import org.pipservices3.commons.errors.ConfigException;
 import org.pipservices3.components.log.CompositeLogger;
 import org.pipservices3.commons.refer.*;
+import org.pipservices3.components.trace.CompositeTracer;
 
 /**
  * Abstract component that supportes configurable dependencies, logging
@@ -24,36 +25,36 @@ import org.pipservices3.commons.refer.*;
  * <ul>
  * <li>*:counters:*:*:1.0       (optional) {@link org.pipservices3.components.count.ICounters} components to pass collected measurements
  * <li>*:logger:*:*:1.0         (optional) {@link org.pipservices3.components.log.ILogger} components to pass log messages
+ * <li>*:tracer:*:*:1.0			(optional) {@link org.pipservices3.components.trace.ITracer} components to trace executed operations
  * <li>...                      References must match configured dependencies.
  * </ul>
  */
 public class Component implements IConfigurable, IReferenceable {
-	
-	protected DependencyResolver _dependencyResolver = new DependencyResolver();
+
+    protected DependencyResolver _dependencyResolver = new DependencyResolver();
     protected CompositeLogger _logger = new CompositeLogger();
     protected CompositeCounters _counters = new CompositeCounters();
+    protected CompositeTracer _tracer = new CompositeTracer();
 
     /**
      * Configures component by passing configuration parameters.
-     * 
-     * @param config    configuration parameters to be set.
+     *
+     * @param config configuration parameters to be set.
      */
-	public void configure(ConfigParams config) throws ConfigException {
-		_dependencyResolver.configure(config);
-        _logger.configure(config);	
-	}
-    
-	/**
-	 * Sets references to dependent components.
-	 * 
-	 * @param references 	references to locate the component dependencies. 
+    public void configure(ConfigParams config) throws ConfigException {
+        _dependencyResolver.configure(config);
+        _logger.configure(config);
+    }
+
+    /**
+     * Sets references to dependent components.
+     *
+     * @param references references to locate the component dependencies.
      */
-	public void setReferences(IReferences references) throws ReferenceException {
-		_dependencyResolver.setReferences(references);
+    public void setReferences(IReferences references) throws ReferenceException {
+        _dependencyResolver.setReferences(references);
         _logger.setReferences(references);
-        _counters.setReferences(references);	
-	}
-
-	
-
+        _counters.setReferences(references);
+        _tracer.setReferences(references);
+    }
 }