+++ src/org/pipservices3/components/trace/ITracer.java
@@ -1,0 +1,12 @@
+package org.pipservices3.components.trace;
+
+/**
+ * Interface for tracer components that capture operation traces.
+ */
+public interface ITracer {
+    void trace(String correlationId, String component, String operation, Long duration);
+
+    void failure(String correlationId, String component, String operation, Exception error, long duration);
+
+    TraceTiming beginTrace(String correlationId, String component, String operation);
+}