+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Component definitions for Java Changelog
 
+## <a name="3.1.0"></a> 3.1.0 (2018-08-15)
+### Features
+- Update the code to the latest version of pip-services toolkit
+- **Lock** -  distributed lock components
+- **State** - components for managing states
+- **Test** - minimal set of test components to make testing easier
+- **Trace** - tracer components
+
+### Bug Fixes
+- Fixed versions conflicts, update dependencies
+
 ## <a name="3.0.0"></a> 3.0.0 (2018-08-15)
 
 Initial public release